# docs/swagger.yaml
basePath: /
definitions:
  requests.CreateModelReq:
    properties:
      address:
        type: string
      api_styles:
        type: string
      api_version:
        type: string
      cache_price:
        type: integer
      deploy_name:
        type: string
      input_price:
        type: integer
      is_private:
        type: integer
      model_id:
        type: string
      name:
        type: string
      output_price:
        type: integer
      owner_id:
        type: integer
    required:
    - model_id
    - name
    type: object
  requests.CreateUserReq:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      user_name:
        type: string
    required:
    - email
    - password
    - user_name
    type: object
  requests.LoginReq:
    properties:
      mail:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - mail
    - password
    type: object
  requests.ModelIdReq:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  requests.RegisterReq:
    properties:
      mail:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - mail
    - password
    type: object
  requests.SetActiveReq:
    properties:
      active:
        enum:
        - 0
        - 1
        type: integer
      id:
        type: integer
    required:
    - id
    type: object
  requests.UpdateModelReq:
    properties:
      address:
        type: string
      api_version:
        type: string
      deploy_name:
        type: string
      id:
        type: integer
      input_price:
        type: integer
      name:
        type: string
      output_price:
        type: integer
      status:
        type: string
    required:
    - id
    type: object
  requests.UpdateUserReq:
    properties:
      id:
        type: integer
      phone:
        type: string
      real_name:
        type: string
      user_name:
        type: string
    required:
    - id
    type: object
  requests.UserIdReq:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  responses.DefaultResponse:
    type: object
host: localhost:8080
info:
  contact: {}
  description: LLM 管理平台 API 文档
  title: TopModelsPlatform API
  version: "1.0"
paths:
  /model/create:
    post:
      consumes:
      - application/json
      description: 创建新模型
      parameters:
      - description: 创建模型请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateModelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DefaultResponse'
      summary: 创建模型
      tags:
      - Model
  /model/delete:
    post:
      consumes:
      - application/json
      description: 删除指定模型
      parameters:
      - description: 模型ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.ModelIdReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DefaultResponse'
      summary: 删除模型
      tags:
      - Model
  /model/get:
    get:
      consumes:
      - application/json
      description: 根据ID获取模型详情
      parameters:
      - description: 模型ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DefaultResponse'
      summary: 获取模型详情
      tags:
      - Model
  /model/list:
    get:
      consumes:
      - application/json
      description: 分页获取模型列表
      parameters:
      - description: 跳过条数
        in: query
        name: skip
        type: integer
      - description: 每页条数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DefaultResponse'
      summary: 获取模型列表
      tags:
      - Model
  /model/update:
    post:
      consumes:
      - application/json
      description: 更新模型信息
      parameters:
      - description: 更新模型请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateModelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DefaultResponse'
      summary: 更新模型
      tags:
      - Model
  /platform/login:
    post:
      consumes:
      - application/json
      description: 用户登录接口
      parameters:
      - description: 登录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DefaultResponse'
      summary: 用户登录
      tags:
      - Platform
  /platform/register:
    post:
      consumes:
      - application/json
      description: 平台用户注册接口
      parameters:
      - description: 注册请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DefaultResponse'
      summary: 用户注册
      tags:
      - Platform
  /user/create:
    post:
      consumes:
      - application/json
      description: 创建新用户
      parameters:
      - description: 创建用户请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DefaultResponse'
      summary: 创建用户
      tags:
      - User
  /user/delete:
    post:
      consumes:
      - application/json
      description: 软删除用户
      parameters:
      - description: 用户ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.UserIdReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DefaultResponse'
      summary: 删除用户
      tags:
      - User
  /user/get:
    get:
      consumes:
      - application/json
      description: 根据ID获取用户详情
      parameters:
      - description: 用户ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DefaultResponse'
      summary: 获取用户详情
      tags:
      - User
  /user/list:
    get:
      consumes:
      - application/json
      description: 分页获取用户列表
      parameters:
      - description: 跳过条数
        in: query
        name: skip
        type: integer
      - description: 每页条数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DefaultResponse'
      summary: 获取用户列表
      tags:
      - User
  /user/setActive:
    post:
      consumes:
      - application/json
      description: 设置用户激活或禁用
      parameters:
      - description: 设置激活状态请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.SetActiveReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DefaultResponse'
      summary: 设置用户激活状态
      tags:
      - User
  /user/update:
    post:
      consumes:
      - application/json
      description: 更新用户信息
      parameters:
      - description: 更新用户请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DefaultResponse'
      summary: 更新用户
      tags:
      - User
swagger: "2.0"